requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
bitarray==3.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
colorama==0.4.6
cytoolz==1.1.0
defusedxml==0.7.1
Django==6.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-guardian==3.2.0
djangorestframework==3.16.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
hexbytes==1.3.1
httpie==3.2.4
idna==3.11
Markdown==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
packaging==25.0
parsimonious==0.10.0
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-solc-x==2.0.5
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PySocks==1.7.1
python-dotenv==1.2.1
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rlp==4.1.0
setuptools==80.9.0
sqlparse==0.5.4
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
